// src/app/shared/enum/notification/notification-type.enum.ts
export enum NotificationTypeEnum {
    ERREUR_404 = "Erreur 404",
    ATTENTION = "Attention",
    SOCKET = "SOCKET",
    INACTIVITE = "Inactivité",
    NAVIGATEUR_OBSOLETE = "Votre navigateur est obsolète",
    MISE_A_JOUR = "Mise à jour",

    TRANSACTION_EN_COURS = "Transaction en cours",
    TRANSACTION_OK = "Transaction terminée",
    TRANSACTION_NOK = "Transaction non aboutie",
    ERREUR = "Erreur",
    ERREUR_TRAITEMENT = "Erreur de traitement",
    ERREUR_SERVEUR = "Erreur serveur",
    ERREUR_INTERNE = "Erreur interne",
    ERREUR_AUTHENTIFICATION = "Erreur d'authentification",
    ERREUR_CONNEXION = "Erreur de connexion",
    ERREUR_AUTORISATION = "Erreur d'autorisation",
    CONNEXION_NOK = "Connexion échouée",
    MODIF_IMPOSSIBLE = "Modification impossible",
    SIGNATURE_NOK = "Echec de la signature",
    INFO_INCORRECTE = "Informations incorrecte",
    CHGT_SOUSCRIPTEUR_IMPOSSIBLE = "Changement du souscripteur impossible",
    CHGT_IMMATRICULATION_IMPOSSIBLE = "Changement d'immatriculation impossible",
    CREATION_EN_COURS = "Création en cours",
    CREATION_IMPOSSIBLE = "Création impossible",
    CREATION_NOK = "Création échouée",
    SUPPRESSION_IMPOSSIBLE = "Suppression impossible",
    CLOTURE_OK = "Clôture réussie",
    CLOTURE_NOK = "Clôture échouée",
    ACTIVATION_NOK = "Activation échouée",
    ACTIVATION_OK = "Activation réussie",
    EXTRACTION_OK = "Extraction terminée",
    ENCAISSEMENT_MODIF = "Encaissement modifié",
    ENCAISSEMENT_SUPPRIM = "Encaissement supprimé",
    IMPUTATION_OK = "Imputation terminée",
    IMPUTATION_MANUELLE = "Imputation manuelle",
    IMPUTATION_ANNULEE = "Imputation annulée",
    MODIFIER_ETAT_TRANSACTION = "Modification effectuée",
    REINIT_NOK = "Echec de la réinitialisation",
    REINIT = "Réinitialisation",
    REINIT_EN_COURS = "Réinitialisation en cours",
    REINIT_OK = "Réinitialisation réussie",
    UPLOAD_OK = "Upload terminé",
    ABANDON_OK = "Abandon terminé",
    CREATION_NOTE = "Création note",
    MODIF_OK = "Modification terminée",
    CREATION_COMPTE = "Le compte utilisateur a été crée",
    RESIL_OK = "Résiliation terminée",
    SUSPENSION_OK = "Suspension terminée",
    DEMANDE_ENVOYEE = "Demande envoyée",
    DEMANDE_EFFECTUEE = "Demande effectuée",
    VERSION_ANNULE = "Version annulée",
    CHEQUE_AJOUT = "Chèque ajouté",

    VALIDATION = "Validation",
}