// src/app/shared/enum/notification/notification-message.enum.ts
export enum NotificationMessageEnum {
	BLOC_AJOUT = "Votre bloc de timbre a bien été ajouté.",
	BLOC_AJOUT_NOK = "Votre bloc de timbre n'a pas pu être ajouté.",
	BLOC_MODIF = "Votre bloc de timbre a bien été modifié.",
	BLOC_MODIF_NOK = "Votre bloc de timbre n'a pas pu être modifié.",
	TIMBRE_AJOUT = "Votre timbre a bien été ajouté.",
	TIMBRE_AJOUT_NOK = "Votre timbre n'a pas pu être ajouté.",
	TIMBRE_MODIF = "Votre timbre a bien été modifié.",
	TIMBRE_MODIF_NOK = "Votre timbre n'a pas pu être modifié.",
	PAYS_MODIF = "Votre pays a bien été modifié.",
	PAYS_MODIF_NOK = "Votre pays n'a pas pu être modifié.",
	PAYS_AJOUT = "Ajout de timbres effectué.",
	PAYS_AJOUT_NOK = "Ajout de timbres n'a pas pu être effectué.",
}